iOS and Android build

+++ include/sampling.hpp
@@ -31,8 +31,8 @@ parse(shake128::shake128_t& hasher, std::span<field::zq_t, ntt::N> poly)
     hasher.squeeze(buf);
 
     for (size_t off = 0; (off < buf.size()) && (coeff_idx < n); off += 3) {
-      const uint16_t d1 = (static_cast<uint16_t>(buf[off + 1] & 0x0f) << 8) | static_cast<uint16_t>(buf[off + 0]);
-      const uint16_t d2 = (static_cast<uint16_t>(buf[off + 2]) << 4) | (static_cast<uint16_t>(buf[off + 1] >> 4));
+      const uint16_t d1 = static_cast<uint16_t>(static_cast<uint16_t>(buf[off + 1] & 0x0f) << 8) | static_cast<uint16_t>(buf[off + 0]);
+      const uint16_t d2 = static_cast<uint16_t>(static_cast<uint16_t>(buf[off + 2]) << 4) | (static_cast<uint16_t>(buf[off + 1] >> 4));
 
       if (d1 < field::Q) {
         poly[coeff_idx] = field::zq_t(d1);
