+++ include/serialize.hpp
@@ -26,10 +26,10 @@ encode(std::span<const field::zq_t, ntt::N> poly, std::span<uint8_t, 32 * l> arr
 
     for (size_t i = 0; i < itr_cnt; i++) {
       const size_t off = i << 3;
-      arr[i] = (static_cast<uint8_t>(poly[off + 7].raw() & one) << 7) | (static_cast<uint8_t>(poly[off + 6].raw() & one) << 6) |
+      arr[i] = static_cast<uint8_t>((static_cast<uint8_t>(poly[off + 7].raw() & one) << 7) | (static_cast<uint8_t>(poly[off + 6].raw() & one) << 6) |
                (static_cast<uint8_t>(poly[off + 5].raw() & one) << 5) | (static_cast<uint8_t>(poly[off + 4].raw() & one) << 4) |
                (static_cast<uint8_t>(poly[off + 3].raw() & one) << 3) | (static_cast<uint8_t>(poly[off + 2].raw() & one) << 2) |
-               (static_cast<uint8_t>(poly[off + 1].raw() & one) << 1) | (static_cast<uint8_t>(poly[off + 0].raw() & one) << 0);
+               (static_cast<uint8_t>(poly[off + 1].raw() & one) << 1) | (static_cast<uint8_t>(poly[off + 0].raw() & one) << 0));
     }
   } else if constexpr (l == 4) {
     constexpr size_t itr_cnt = ntt::N >> 1;
@@ -37,7 +37,7 @@ encode(std::span<const field::zq_t, ntt::N> poly, std::span<uint8_t, 32 * l> arr
 
     for (size_t i = 0; i < itr_cnt; i++) {
       const size_t off = i << 1;
-      arr[i] = (static_cast<uint8_t>(poly[off + 1].raw() & msk) << 4) | static_cast<uint8_t>(poly[off + 0].raw() & msk);
+      arr[i] = static_cast<uint8_t>((static_cast<uint8_t>(poly[off + 1].raw() & msk) << 4) | static_cast<uint8_t>(poly[off + 0].raw() & msk));
     }
   } else if constexpr (l == 5) {
     constexpr size_t itr_cnt = ntt::N >> 3;
@@ -60,11 +60,11 @@ encode(std::span<const field::zq_t, ntt::N> poly, std::span<uint8_t, 32 * l> arr
       const auto t6 = poly[poff + 6].raw();
       const auto t7 = poly[poff + 7].raw();
 
-      arr[boff + 0] = (static_cast<uint8_t>(t1 & mask3) << 5) | (static_cast<uint8_t>(t0 & mask5) << 0);
-      arr[boff + 1] = (static_cast<uint8_t>(t3 & mask1) << 7) | (static_cast<uint8_t>(t2 & mask5) << 2) | static_cast<uint8_t>((t1 >> 3) & mask2);
-      arr[boff + 2] = (static_cast<uint8_t>(t4 & mask4) << 4) | static_cast<uint8_t>((t3 >> 1) & mask4);
-      arr[boff + 3] = (static_cast<uint8_t>(t6 & mask2) << 6) | (static_cast<uint8_t>(t5 & mask5) << 1) | static_cast<uint8_t>((t4 >> 4) & mask1);
-      arr[boff + 4] = (static_cast<uint8_t>(t7 & mask5) << 3) | static_cast<uint8_t>((t6 >> 2) & mask3);
+      arr[boff + 0] = static_cast<uint8_t>((static_cast<uint8_t>(t1 & mask3) << 5) | (static_cast<uint8_t>(t0 & mask5) << 0));
+      arr[boff + 1] = static_cast<uint8_t>((static_cast<uint8_t>(t3 & mask1) << 7) | (static_cast<uint8_t>(t2 & mask5) << 2) | static_cast<uint8_t>((t1 >> 3) & mask2));
+      arr[boff + 2] = static_cast<uint8_t>((static_cast<uint8_t>(t4 & mask4) << 4) | static_cast<uint8_t>((t3 >> 1) & mask4));
+      arr[boff + 3] = static_cast<uint8_t>((static_cast<uint8_t>(t6 & mask2) << 6) | (static_cast<uint8_t>(t5 & mask5) << 1) | static_cast<uint8_t>((t4 >> 4) & mask1));
+      arr[boff + 4] = static_cast<uint8_t>((static_cast<uint8_t>(t7 & mask5) << 3) | static_cast<uint8_t>((t6 >> 2) & mask3));
     }
   } else if constexpr (l == 10) {
     constexpr size_t itr_cnt = ntt::N >> 2;
@@ -204,12 +204,12 @@ decode(std::span<const uint8_t, 32 * l> arr, std::span<field::zq_t, ntt::N> poly
       const auto t7 = static_cast<uint16_t>((arr[boff + 4] >> 3) & mask5);
 
       poly[poff + 0] = field::zq_t(t0);
-      poly[poff + 1] = field::zq_t(t1);
+      poly[poff + 1] = field::zq_t(static_cast<uint16_t>(t1));
       poly[poff + 2] = field::zq_t(t2);
-      poly[poff + 3] = field::zq_t(t3);
-      poly[poff + 4] = field::zq_t(t4);
+      poly[poff + 3] = field::zq_t(static_cast<uint16_t>(t3));
+      poly[poff + 4] = field::zq_t(static_cast<uint16_t>(t4));
       poly[poff + 5] = field::zq_t(t5);
-      poly[poff + 6] = field::zq_t(t6);
+      poly[poff + 6] = field::zq_t(static_cast<uint16_t>(t6));
       poly[poff + 7] = field::zq_t(t7);
     }
   } else if constexpr (l == 10) {
@@ -227,10 +227,10 @@ decode(std::span<const uint8_t, 32 * l> arr, std::span<field::zq_t, ntt::N> poly
       const auto t2 = (static_cast<uint16_t>(arr[boff + 3] & mask6) << 4) | static_cast<uint16_t>(arr[boff + 2] >> 4);
       const auto t3 = (static_cast<uint16_t>(arr[boff + 4]) << 2) | static_cast<uint16_t>(arr[boff + 3] >> 6);
 
-      poly[poff + 0] = field::zq_t(t0);
-      poly[poff + 1] = field::zq_t(t1);
-      poly[poff + 2] = field::zq_t(t2);
-      poly[poff + 3] = field::zq_t(t3);
+      poly[poff + 0] = field::zq_t(static_cast<uint16_t>(t0));
+      poly[poff + 1] = field::zq_t(static_cast<uint16_t>(t1));
+      poly[poff + 2] = field::zq_t(static_cast<uint16_t>(t2));
+      poly[poff + 3] = field::zq_t(static_cast<uint16_t>(t3));
     }
   } else if constexpr (l == 11) {
     constexpr size_t itr_cnt = ntt::N >> 3;
@@ -255,14 +255,14 @@ decode(std::span<const uint8_t, 32 * l> arr, std::span<field::zq_t, ntt::N> poly
       const auto t6 = (static_cast<uint16_t>(arr[boff + 9] & mask5) << 6) | static_cast<uint16_t>(arr[boff + 8] >> 2);
       const auto t7 = (static_cast<uint16_t>(arr[boff + 10]) << 3) | static_cast<uint16_t>(arr[boff + 9] >> 5);
 
-      poly[poff + 0] = field::zq_t(t0);
-      poly[poff + 1] = field::zq_t(t1);
-      poly[poff + 2] = field::zq_t(t2);
-      poly[poff + 3] = field::zq_t(t3);
-      poly[poff + 4] = field::zq_t(t4);
-      poly[poff + 5] = field::zq_t(t5);
-      poly[poff + 6] = field::zq_t(t6);
-      poly[poff + 7] = field::zq_t(t7);
+      poly[poff + 0] = field::zq_t(static_cast<uint16_t>(t0));
+      poly[poff + 1] = field::zq_t(static_cast<uint16_t>(t1));
+      poly[poff + 2] = field::zq_t(static_cast<uint16_t>(t2));
+      poly[poff + 3] = field::zq_t(static_cast<uint16_t>(t3));
+      poly[poff + 4] = field::zq_t(static_cast<uint16_t>(t4));
+      poly[poff + 5] = field::zq_t(static_cast<uint16_t>(t5));
+      poly[poff + 6] = field::zq_t(static_cast<uint16_t>(t6));
+      poly[poff + 7] = field::zq_t(static_cast<uint16_t>(t7));
     }
   } else {
     static_assert(l == 12, "l must be equal to 12 !");
@@ -277,8 +277,8 @@ decode(std::span<const uint8_t, 32 * l> arr, std::span<field::zq_t, ntt::N> poly
       const auto t0 = (static_cast<uint16_t>(arr[boff + 1] & mask4) << 8) | static_cast<uint16_t>(arr[boff + 0]);
       const auto t1 = (static_cast<uint16_t>(arr[boff + 2]) << 4) | static_cast<uint16_t>(arr[boff + 1] >> 4);
 
-      poly[poff + 0] = field::zq_t(t0);
-      poly[poff + 1] = field::zq_t(t1);
+      poly[poff + 0] = field::zq_t(static_cast<uint16_t>(t0));
+      poly[poff + 1] = field::zq_t(static_cast<uint16_t>(t1));
     }
   }
 }
